+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -7,6 +7,9 @@ import java.io.OutputStream;
 import java.net.Socket;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.request.Request;
+import org.apache.coyote.http11.request.StartLine;
+import org.apache.coyote.http11.response.ResponseProcessor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
