+++ tomcat/src/main/java/org/apache/coyote/http11/response/ResponseHeaders.java
@@ -3,8 +3,6 @@ package org.apache.coyote.http11.response;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.stream.Collectors;
-import org.apache.coyote.http11.ContentType;
-import org.apache.coyote.http11.ResponseEntity;
 
 public class ResponseHeaders {
