+++ tomcat/src/test/java/org/apache/coyote/http11/ControllerTest.java
@@ -8,6 +8,10 @@ import java.net.URISyntaxException;
 import nextstep.jwp.exception.AuthenticationException;
 import nextstep.jwp.exception.ResourceNotFoundException;
 import nextstep.jwp.exception.UserNotFoundException;
+import org.apache.coyote.http11.request.Request;
+import org.apache.coyote.http11.request.StartLine;
+import org.apache.coyote.http11.response.HttpStatus;
+import org.apache.coyote.http11.response.ResponseEntity;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
