+++ tomcat/src/test/java/org/apache/coyote/http11/response/GeneralTest.java
@@ -3,8 +3,6 @@ package org.apache.coyote.http11.response;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.stream.Stream;
-import org.apache.coyote.http11.HttpStatus;
-import org.apache.coyote.http11.response.General;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
