+++ tomcat/src/test/java/org/apache/coyote/http11/response/ResponseBodyTest.java
@@ -8,10 +8,6 @@ import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import org.apache.coyote.http11.HttpStatus;
-import org.apache.coyote.http11.ResponseEntity;
-import org.apache.coyote.http11.response.ResponseBody;
-import org.apache.coyote.http11.response.ResponseHeaders;
 import org.junit.jupiter.api.Test;
 
 class ResponseBodyTest {
