+++ tomcat/src/test/java/org/apache/coyote/http11/response/ResponseHeadersTest.java
@@ -3,9 +3,6 @@ package org.apache.coyote.http11.response;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.stream.Stream;
-import org.apache.coyote.http11.HttpStatus;
-import org.apache.coyote.http11.ResponseEntity;
-import org.apache.coyote.http11.response.ResponseHeaders;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
